+++ src/components/icons/PointMeasurementIcon.vue
@@ -1,0 +1,36 @@
+<!-- eslint-disable max-len -->
+
+<template>
+  <svg
+    id="icon-point-coordinates"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 24 24"
+  >
+    <path
+      id="Pfad_716"
+      d="m1.67,18.17v-1.7h4.77v1.7l-1.95,2.47v.15h1.95v1.7H1.67v-1.62l2-2.55v-.15h-2Z"
+      fill="currentColor"
+      stroke-width="0"
+    />
+    <path
+      id="Pfad_715"
+      d="m20.05,16.47l.55,1.65.7-1.65h2.05l-1.3,2.87,1.22,3.16h-2.14l-.53-1.6-.55,1.6h-2.12l1.29-2.89-1.29-3.14h2.12Z"
+      fill="currentColor"
+      stroke-width="0"
+    />
+    <path
+      id="Pfad_717"
+      d="m13.14,7.5h-2.09v-1.89l-1.72-4.14h2.22l.57,1.77.58-1.77h2.21l-1.77,4.14v1.89Z"
+      fill="currentColor"
+      stroke-width="0"
+    />
+    <circle
+      id="Ellipse_41"
+      cx="12"
+      cy="13.47"
+      r="3"
+      fill="currentColor"
+      stroke-width="0"
+    />
+  </svg>
+</template>